re-display minimized menu state when opening anchors
When in a minimized state, a button to open up the sidemenu appears to
help easily navigate to a desired page. The original implementation
assumed that the request to navigate would be enough to restore the
user's state as it a new page request would be performed and a clean
minimized state would appear. However, if the toctree provides a
reference to a local target, the user will jump to the target but the
sidemenu will still remain.

The following tweaks the menu call to automatically restore the
minimize state if any link is clicked.

Signed-off-by: James Knight <james.d.knight@live.com>