force register sphinx_alice_theme extension
When the sphinx-alice-theme is installed via pip, the entry point will
allow the theme to be properly registered into the Sphinx application;
however, when locally building/developing in a Git environment, there is
no entry-point to register the extension and the `setup` phase cannot be
invoked. Help force this processing of this theme's implementation by
forcefully registering this extension.

Signed-off-by: James Knight <james.d.knight@live.com>